+++ src/modules/grid.module.ts
@@ -3,7 +3,7 @@ import {CommonModule as AngularCommonModule} from "@angular/common";
 import {CommonModule} from "@anglr/common";
 
 import {GridComponent} from "../components/grid/grid.component";
-import {VirtualScrollPagingComponent, PreviousNextPagingComponent, LoadMorePagingComponent, BasicPagingComponent, NoPagingComponent, QueryPagingInitializerComponent, NoPagingInitializerComponent} from "../plugins/paging";
+import {PageVirtualScrollPagingComponent, ContentVirtualScrollPagingComponent, PreviousNextPagingComponent, LoadMorePagingComponent, BasicPagingComponent, NoPagingComponent, QueryPagingInitializerComponent, NoPagingInitializerComponent} from "../plugins/paging";
 import {SyncDataLoaderComponent, AsyncDataLoaderComponent} from "../plugins/dataLoader";
 import {BasicTableColumnComponent, BasicTableMetadataGathererComponent} from "../components/metadata";
 import {NoMetadataSelectorComponent, AdvancedMetadataSelectorComponent} from "../plugins/metadataSelector";
@@ -26,7 +26,8 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
         BasicTableMetadataGathererComponent,
         BasicPagingComponent,
         LoadMorePagingComponent,
-        VirtualScrollPagingComponent,
+        ContentVirtualScrollPagingComponent,
+        PageVirtualScrollPagingComponent,
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryPagingInitializerComponent,
@@ -50,7 +51,8 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
     [
         BasicPagingComponent,
         LoadMorePagingComponent,
-        VirtualScrollPagingComponent,
+        ContentVirtualScrollPagingComponent,
+        PageVirtualScrollPagingComponent,
         PreviousNextPagingComponent,
         NoPagingComponent,
         QueryPagingInitializerComponent,
@@ -77,7 +79,8 @@ import {BasicRowSelectorComponent} from "../plugins/rowSelector";
         BasicTableMetadataGathererComponent,
         BasicPagingComponent,
         LoadMorePagingComponent,
-        VirtualScrollPagingComponent,
+        ContentVirtualScrollPagingComponent,
+        PageVirtualScrollPagingComponent,
         PreviousNextPagingComponent
     ],
     providers: