+++ src/plugins/paging/virtualScroll/pageVirtualScroll/pageVirtualScrollPaging.interface.ts
@@ -1,0 +1,15 @@
+import {VirtualScrollPagingOptions, VirtualScrollPaging} from "../virtualScrollPaging.interface";
+
+/**
+ * Page virtual scroll paging options
+ */
+export interface PageVirtualScrollPagingOptions extends VirtualScrollPagingOptions
+{
+}
+
+/**
+ * Public API for 'PageVirtualScrollPaging'
+ */
+export interface PageVirtualScrollPaging extends VirtualScrollPaging
+{
+}