+++ src/plugins/paging/virtualScroll/virtualScrollPaging.interface.ts
@@ -16,15 +16,10 @@ export interface VirtualScrollPagingOptions extends PagingOptions<CssClassesVirt
      * Offset between 0 and 1 which indicates when to start loading next page, 0.8 means 80% of scrolling distance
      */
     loadOffsetTreshold?: number;
-
-    /**
-     * Css max height that is applied to table
-     */
-    maxHeight?: string;
 }
 
 /**
- * Public API for 'VirtualScrollPagingComponent'
+ * Public API for 'VirtualScrollPaging'
  */
 export interface VirtualScrollPaging extends Paging
 {